+++ src/Attribute/TypeScriptInterface.php
@@ -1,7 +1,15 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Paneon\PhpToTypeScript\Attribute;
+
+use Attribute;
+
+/**
+ * Marks a class for TypeScript interface generation.
+ *
+ * @deprecated Use \Paneon\PhpToTypeScript\Attribute\TypeScript instead
+ */
+#[Attribute(flags: Attribute::TARGET_CLASS)]
 class TypeScriptInterface
 {
-
 }